Implemented sync task execution.
Note:
Performance is very bad, because CPU usage is very high since we are not sleeping in Scheduler::idle_sleep().
This might be solved by waiting on mutex when there is no work to do, but in that approach we will need to signal for every event for thread to wake up (when new task arrives, when m_event is signaled, when I/O completes etc.).