[Bugfix] Fixing infinite loop when running in Release
Release build has bug which manifest in inifinite loop. This is because
has_tasks doesn't lock before calling empty(), hance what probably
happens is that that call is optimized so it always returns false.
Adding a mutex tells optimizer that m_tasks will be access by other
threads and that it shouldn't thouch it :)