Implemented basic I/O functionality (for windows OS).
This implementation assumes async I/O capabilities on the OS level.
When I/O is issued, it calls one of async I/O functions, and worker will be blocked until I/O request is done. It will park itself in the io_pending queue. Every time we are entering scheduler we will check whether I/O request is completed and move worker to runnable queue if it is.
So, from user perspective I/O request will be executed synchronously since worker is blocked, but we are executing asynchronous I/O on the OS level.

For the next improvements, we should provide async I/O functionality from the user perspective and also support async I/O on OSs that do not support async I/O calls.